added async

+++ demo-api/src/user/user.controller.ts
@@ -10,22 +10,22 @@ export class UserController {
     constructor(private readonly userService: UserService) { }
 
     @Get()
-    getAllUsers(): Promise<User[]> {
-        return this.userService.findAllUsers();
+    async getAllUsers(): Promise<User[]> {
+        return await this.userService.findAllUsers();
     }
 
     @Get(':id')
-    getUserById(@Param('id') id: string): Promise<User> {
-        return this.userService.findOneUser(id);
+    async getUserById(@Param('id') id: string): Promise<User> {
+        return await this.userService.findOneUser(id);
     }
 
     @Post()
-    create(@Body() createDto: CreateDto): Promise<User> {
+    async create(@Body() createDto: CreateDto): Promise<User> {
         const user = new User();
         user.firstName = createDto.firstName;
         user.lastName = createDto.lastName;
         user.isActive = createDto.isActive;
-        return this.userService.createUser(user);
+        return await this.userService.createUser(user);
     }
 
     @Put(':id')
@@ -40,11 +40,15 @@ export class UserController {
         user.lastName = updateDto.lastName;
         user.isActive = updateDto.isActive; 
 
-        return this.userService.updateUser(user);;
+        return await this.userService.updateUser(user);;
     }
 
     @Delete(':id')
-    remove(@Param('id') id: string): Promise<void> {
-        return this.userService.removeUser(id);
+    async remove(@Param('id') id: string): Promise<void> {
+        const user = await this.userService.findOneUser(id);
+        if(user === undefined){
+            throw new NotFoundException(`User with id ${id} not found`);
+        }
+        return await this.userService.removeUser(id);
     }
 }
